requirements.txt
python-dateutil>=2.8.1
plotly>=4.14.3
pytz>=2021.1